# perf: improve performance for read coalescing