@nsmith- 2024-10-23
After the change, is the read coalescing algorithm time negligble?