@nikoladze 2024-10-23
> After the change, is the read coalescing algorithm time negligble?

yes, `_merge_adjacent` for the 50k ranges in the example now takes in the order of 30ms