@jpivarski 2024-10-23
Although I'd consider a _O(n²)_ → _O(n)_ fix to be a serious bug-fix, I'm labeling this commit as "perf" for clarity, to indicate that user-visible behavior doesn't change.